Extract auth and error core services
